core/library/requirements.txt
beautifulsoup4
certifi
scikit-build
cmake
colorama
Django
Flask
gitpython
hackingtools==3.0.0.034
html5lib
Jinja2
keyring
lxml
mechanize
pandas
paramiko
Pillow
progressbar2
Pycco
Pygments
PyInstaller
pylint
PyPDF2
pyngrok
pyreadline
python-nmap
requests
requests_html
scp
shodan
six
python-slugify
pysia
twine
typed-ast
urllib3
virustotal-python
xlsxwriter